+++ templates/index.html
@@ -34,20 +34,6 @@
     <div class="captions" id="captions">
       <span>Realtime speech transcription API</span>
     </div>
-    <div class="button-container">
-      <a
-        href="https://console.deepgram.com/signup"
-        class="info-button sign-up"
-        target="_blank"
-        >Sign Up</a
-      >
-      <a
-        href="https://developers.deepgram.com/docs/introduction"
-        class="info-button docs"
-        target="_blank"
-        >Read the Docs</a
-      >
-    </div>
     <script src="../static/script.js"></script>
   </body>
 </html>
